// src/enums/data/ID/GridID.ts
/** A list of unique grid entities, indexed using their GridEntityID. */
export enum GridID {
  NULL = "0.0",
  DECORATION = "1.0", // Check if more.
  ROCK = "2.0",
  EVENT_ROCK = "2.1",
  BLOCK = "3.0",
  ROCK_TINTED = "4.0",
  ROCK_BOMB = "5.0",
  ROCK_ALT = "6.0",
  PIT = "7.0",
  PIT_FISSURE_SPAWNER = "7.16",
  SPIKES = "8.0",
  SPIKES_ON_OFF = "9.0",
  SPIDER_WEB = "10.0",
  LOCK = "11.0",
  TNT = "12.0",
  FIREPLACE = "13.0",
  FIREPLACE_RED = "13.1",
  POOP = "14.0",
  POOP_RED = "14.1",
  POOP_CORNY = "14.2",
  POOP_GOLDEN = "14.3",
  POOP_RAINBOW = "14.4",
  POOP_BLACK = "14.5",
  POOP_WHITE = "14.6",
  POOP_GIANT_TOP_LEFT = "14.7",
  POOP_GIANT_TOP_RIGHT = "14.8",
  POOP_GIANT_BOTTOM_LEFT = "14.9",
  POOP_GIANT_BOTTOM_RIGHT = "14.10",
  POOP_CHARMING = "14.11",
  WALL = "15.0",
  DOOR = "16.0",
  DOOR_LOCKED = "16.1",
  DOOR_LOCKED_DOUBLE = "16.2",
  DOOR_LOCKED_CRACKED = "16.3",
  DOOR_LOCKED_BARRED = "16.4",
  DOOR_LOCKED_KEY_FAMILIAR = "16.5",
  DOOR_LOCKED_GREED = "16.6",
  DOOR_HIDDEN = "16.7",
  DOOR_UNLOCKED = "16.8",
  TRAPDOOR = "17.0",
  VOID_PORTAL = "17.1",
  CRAWL_SPACE = "18.0",
  CRAWL_SPACE_GREAT_GIDEON = "18.1",
  CRAWL_SPACE_SECRET_SHOP = "18.2",
  CRAWL_SPACE_PASSAGE_TO_BEGINNING_OF_FLOOR = "18.3",
  CRAWL_SPACE_NULL = "18.4",
  GRAVITY = "19.0",
  PRESSURE_PLATE = "20.0",
  REWARD_PLATE = "20.1",
  GREED_PLATE = "20.2",
  RAIL_PLATE = "20.3",
  KILL_ALL_ENEMIES_PLATE = "20.9",
  SPAWN_ROCKS_PLATE = "20.10",
  STATUE_DEVIL = "21.0",
  STATUE_ANGEL = "21.1",
  ROCK_SUPER_SPECIAL = "22.0",
  TELEPORTER = "23.0",
  PILLAR = "24.0",
  ROCK_SPIKED = "25.0",
  ROCK_FOOL_CARD = "26.0",
  ROCK_GOLD = "27.0",
}